Restore mHackerSwapChain field in HackerDevice.
HackerDevice->GetHackerSwapChain is used in a couple of
places, where the backbuffer is needed.  This restores the
GetHackerSwapChain so they can fetch swap chain as
necessary.

Sets the reference in the HackerSwapChain constructor,
because the input is HackerDevice and HackerContext.

Since we can now be certain there is only one path to
constructing a HackerSwapChain, this will be reliable.