Restore HackerUpscalingSwapChain.
This restores the code to the HookedDXGI CreateSwapChain
sequence, and creates the HackerUpscalingSwapChain in the
case where it is set in d3dx.ini.

Also did some code cleanup to make this match our code
style better.

Since so many changes were made during refactoring, this
is fairly unlikely to run properly.  Code compiles, but is
untested.