Restore hooking in HackerDevice/Context.
This revives the temporarily removed hooking feature, but
only does the obvious and simplistic renaming of the
objects to ID3D11Device1 and ID3D11DeviceContext1.

This is incomplete however.  The *1 objects have more functions that need to be copied into the tables and hooked up.  For any game that doesn't care about *1 objects it should still work, as the missing calls would never be called.

Compiles and runs, but is not tested.