Make CreateSwapChain path work with HookedDevice
When we have hooking enabled, the device passed into CreateSwapChain()
will be the original DirectX device with hooks to call into our code,
whereas without hooking it will be the HackerDevice we handed to the
game.

This adds the code to look up the corresponding HackerDevice for the
given DirectX device and use it when creating HackerSwapChain.

This change makes MGSV completely functional with the 1.3 branch :)