Improve error handling in replacement CreateDeviceAndSwapChain.
Since we are breaking the semantics of this call with our
replacement, let's make this more solid, and more clear.

If we get an unexpected scenario, we will now cause a fatal
exit with DoubleBeepExit, which will log and let us know. We
do not ever expect to see this happen, but it's better to have
a hard failure.

Also improves the error handling to more closely follow
what DX would normally do.

Tested to work correctly in Headlander, which calls CreateDeviceAndSwapChain.