Set platform_update=1 by default in d3dx.ini
For this branch, the goal is to not force failures in DX11 calls,
and thus force games to use a downlevel code path.

This changes the default in d3dx.ini to allow all calls through
without forced errors, which will be whatever the game
would normally expect.

Not certain this is the right approach, but we can change
it back if we find it does not work well.  Initial testing has
shown it to be successful, including some of the more
bizarre games.