Move SetFullScreenState into Upscaling case.
If we are not in Upscaling mode, we don't want to be setting
the screen state directly.  That will be handled by ForceDisplayParams
setting the windowed mode or not, as part of the CreateSwapChain.

This moves the SetFullScreenState into the Upscaling case,
so that it only triggers when in that mode.

This was found because of a break while using debug build.