Release Device/Context on CreateSwapChain failure.
In CreateDeviceAndSwapChain, it's theoretically possible to
succeed with Device, but fail with SwapChain, so this change
checks for failure, and releases the Device and Context, then
returns the error to the game, for its normal handling.

Unlikely to ever be seen, but this hews more closely to the
original call semantics.