Upscaling: Fix refcounting on mFakeBackBuffer
GetBuffer() needs to bump the refcount on the buffer it returns - if we
call through to DirectX it will do that for us, but if we return our own
resource we need to do it ourselves.

Should fix a crash ColAngel noted when the command list resource pool
attempted to release a cached resource.

Untested.