Hooking: Remove crash workarounds from hook=recommended
The updated Deviare library fixes these crashes, so we no longer need
the workarounds enabled.

This is especially important since 1.2.68 changed the method we use to
prevent the game from unbinding StereoParams & IniParams to depend on
the XXSetShaderResources calls, so enabling hooking on certain games
could result in 3DMigoto appearing to be mostly non-functional.

It also means that the frame analysis log will now be useful to trace
resources on games using hooking.

The "except_" options are left in for now, just in case it turns out we
still need any of them, but they are no longer selected by
"recommended", and as such should be disabled for anyone updating
3DMigoto unless they explicitly specified them.