Upscaling: Fix GetBuffer not validating riid, and incorrect behaviour on ppSurface=NULL
There were several issues with this function:

The first was that when I added the check on ppSurface I assumed the
code would fall through to the original GetBuffer() call and that would
validate the parameters for us, but that was incorrect since
mFakeSwapChain would be NULL, so it would actually fall through to the
assert(hr) case.

The second issue is that assert(hr) is a noop in release builds, so if
we hit that case it would return S_OK, but have left ppSurface invalid
risking the game crashing if it dereferences it - and that's the best
case scenario. Coupled with the first problem, this is no longer
theoretical.

The third issue is that the riid argument is not validated when using
mFakeBackBuffer upscaling mode and we always return an ID3D11Texture2D
interface, just hoping that it will be castable to whatever interface
was actually requested (which is probably fine in most cases, unless an
unrelated interface was requested).

To fix all of these problems, swap to calling QueryInterface on the
mFakeBackBuffer - that will ensure that the riid is validated as a
supported interface, and should return an error if it or ppSurface is
invalid, and since QueryInterface performs an implicit AddRef, we no
longer need to do this explicitly.

To protect against the second problem ever occurring again, we switch
the assert to use a LogInfo & DoubleBeepExit, which is more consistent
with other hard exits we have used elsewhere in the code. Also make the
same change in the upscaling ResizeBuffers.