Fix refresh rate overrides with UE4 and other games
The refresh rate overrides only worked with games that created a full
screen swap chain from the get go (or when we were forcing that), but
not with games that create a windowed mode swap chain and later change
it to fullscreen - which is the method MSDN recommends, and as such is
very common.

These games set the refresh rate in the ResizeTarget() call, but we only
forced the 2x width override for direct mode in that call. Split out the
code that forces the display mode from ForceDisplayParams() into a
separate call that we also call from ResizeTarget() to solve this.

Currently, this will only force the refresh rate if the swap chain is
already full-screen by the time ResizeTarget() is called. I don't have a
particularly good reason for doing this, other than that's what 3DMigoto
already did when the swap chain was created and maybe there was some
reason for doing so? This is enough to get UE4 games to work with 3D
TV Play since they call ResizeTarget() after SetFullscreenState(), but
MSDN indicates that these calls can be made in any order, and in fact
recommends* doing the opposite to avoid an extra WM_SIZE, so it is likely
that there are other games that will need a little more work, but I want
to see one of these in action before deciding on the best way to handle
that.

* See DirectX Graphics Infrastructure (DXGI): Best Practices:
https://msdn.microsoft.com/en-us/library/windows/desktop/ee417025(v=vs.85).aspx

As an added bonus, this (somehow) avoids the 3D TV Play green flicker
bug introduced in 376.33 in certain games such as Tekken 7.