Fuzzy TextureOverride matching: Assume usage=default
There are some reported driver crashes when forcing the stereomode of
certain resources, and although it is not yet confirmed exactly what
causes this, forcing the creation mode of usage=immutable resources to
mono is one of the possible suspects - even though that should have no
effect, since immutable resources are mono by definition.

To try to minimise the risk of these crashes, this changes the fuzzy
matching to assume match_usage=default by default, and other usage types
would have to be explicitly specified if desired, or the old hash
matching method used instead.

This does carry the risk of breaking fixes using this feature, however
the risk is minimal since it is brand new and so far only used to
replace driver heuristics, which only make sense on usage=default
anyway - all other usage types are inherently mono since they come from
or go to the CPU.

The fuzzy matching could also be used for other reasons, such as texture
filtering or checktextureoverride, but in general we expect those to
mostly be matched by hash, and in many cases will probably be using
usage=default anyway, so the risk of breaking them is still low.