Add workaround for building DirectXTK on SDK 8.0.
With the older version of DirectXMath, the XM_CALLCONV
is not defined.  In order to use DirectXMath with SDK 8.0
we need to add the specified adapter code.

https://github.com/Microsoft/DirectXMath/wiki

This is added in Overlay where we use it directly, and also
in pch.h, which is part of DirectXTK.  DirectXTK needs to
build as well, because it creates a .lib file we link against.