ShaderRegex: Fix command lists not being applied if no ShaderOverride sections exist
Command lists in ShaderRegex sections are ignored in release mode if no
other ShaderOverride sections exist.

This is due to an optimisation path that doesn't update the current
shader hash if it doesn't think we need it, causing the ShaderRegex code
to link the command lists in to a ShaderOverride without an associated
hash.

This is the second time this same optimisation has caught me out now,
and since the problems it causes are not visible in hunting mode they
are easy to miss until an end user complains.

Reported-By: DHR