Change DirectXTK vs2017 project to match project.
This makes the equivalent changes from master/vs2015 to the
DirectXTK vs2017 project file:

Setting to MultiThreaded and MultiThreadedDebug, from the
standard MultiThreadedDLL.  We are static linking this code.
This will be picked up for any 3Dmigoto build, because it
is included as part of the primary configuration.

Also set Exception handling to EHa and BufferSecurityCheck
to match DirectX11 project.

This builds in all 4 variants of x32, x64, debug, release.