Add lookup map to find HackerDevice from an IUnknown
UE4 gets an IDXGIDevice interface from ID3D11Device via QueryInterface.
It than passes this interface to CreateSwapChain, and since we no longer
wrap IDXGIDevice interfaces we don't know what to do with it, illegally
cast it into a HackerDevice and ultimately crash later when we try to
use it to create or use the overlay.

While annoying for us, this is a valid use case since CreateSwapChain
takes an interface to a D3D Device, and an IDXGIDevice counts as one of
those as it is in fact another interface to the same D3D device as the
ID3D11Device is also an interface to - this can easily be demonstrated
by calling QueryInterface(IID_IUnknown) on both and observing that the
returned IUnknown is identical - this is a requirement of COM that all
interfaces to the same object return consistent IUnknown pointers from
QueryInterface (though obviously casting an interface to IUnknown with
C/C++ style casting will not respect this).

We would rather avoid re-adding the IDXGIDevice wrapper and all the
complexity it brings with it, so instead, we add a lookup map that finds
a HackerDevice from any IUnknown. We check if it is already a
HackerDevice via QueryInterface(IID_HackerDevice), and otherwise use
QueryInterface to get the unique IUnknown pointer for the object
instance and use that to find our HackerDevice. Using the IUnknown
pointer means that we will find our HackerDevice regardless of whether
we were given an ID3D11Device or IDXGIDevice.

All HackerDevices will register themselves with this map when they are
created, and unregister themselves when (if) they are destroyed. This
includes devices using hooking, so this map removes the need to lookup
hooked devices separately (the map for hooked devices is left in, since
the key for that is the ID3D11Device*, not the IUnknown*).

We also store this IUnknown pointer in the HackerDevice, because when
the device is released and we go to remove it from the map, we still
need to know the real IUnknown pointer, but we will no longer be able to
look it up with QueryInterface, since the device has already been freed.