Bump Windows SDK Version to 10.0.17763.0
Updating to the latest VS2017 no longer builds, with this error
(specifically during DirectXTK compilation):

1>c:\program files (x86)\windows kits\10\include\10.0.16299.0\winrt\wrl\event.h(316): error C7510: 'Callback': use of dependent template name must be prefixed with 'template'
1>c:\program files (x86)\windows kits\10\include\10.0.16299.0\winrt\wrl\event.h(324): error C7510: 'Callback': use of dependent template name must be prefixed with 'template'

Good job Microsoft :eyeroll. We're not married to this SDK, so just
update to the latest to resolve this issue. This commit is a result of
the following command to bump the Windows SDK up to 10.0.17763.0:

sed -i 's/<WindowsTargetPlatformVersion>.*<\/WindowsTargetPlatformVersion>/<WindowsTargetPlatformVersion>10.0.17763.0<\/WindowsTargetPlatformVersion>/' $(find . -iname '*.vcxproj')