IUnknown analysis: Backport to Win 8.1 SDK with reduced functionality
This checks which versions have been defined by the SDK in use to
determine which headers it may include, and which IIDs the SDK knows
about. This allows it to be used with the Windows 8.1 and 10.0.14393.0
SDKs with reduced functionality (in particular, DX12 interfaces will not
be checked without the Windows 10 SDK).