IUnknown analysis: Backport to Win 8.0 SDK with reduced functionality
This continues the backport from the previous commit to skip headers and
interfaces that require the 8.1 SDK or later, allowing this to be
included in our vs2013 branch.

This is in a separate commit because switching between 8.1 and more
recent SDKs is just a few clicks on the vs2015 branch, but switching
back to the 8.0 SDK is not possible on that branch and requires a legacy
branch that has not yet been forward ported to the newer SDKs.