IUnknown Analysis: Properly include SDK version header
This header is implicitly included by the various DX headers, but needs
to be included explicitly before testing the versions it defines - this
of course worked before backporting when we didn't have any of those
tests, and worked during backporting since the #ifdefs were expected to
be skipped either way, but of course broke it on the vs2015 branch with
newer SDKs as the tests would be inconsistent throughout the file.

Explicitly include the version header to work with any SDK version.