Swap Chain: Do not treat unknown devices as fatal
CreateSwapChain can be called with an ID3D12CommandQueue on Windows 10,
or possibly an ID3D10Device for old DX10 only games. Currently, we
DoubleBeepExit() when this happens, which makes some sense because
3DMigoto is not going to be able to do much in these cases, but there
could be some cases where we might want to allow the swap chain to be
created anyway - perhaps the application creates several swap chains
(maybe an overlay), or perhaps we would rather render an error to the
DX12 swap chain informing the user to switch the game to use DX11
(though we'd need to migrate to the vs2015 branch to even consider doing
so).

Either way, let's try to allow the CreateSwapChain functions to succeed
in these cases - we might decide to go back to DoubleBeepExit() as the
best strategy, but let's keep our options open.

Refactors CreateSwapChain and CreateSwapChainForHwnd to use common code
to find the HackerDevice, and to not fail in the event the HackerDevice
is not found.

We still treat the case where a HackerDevice cannot be found as fatal if
the IUnknown supports the ID3D11Device interface - if this happens, we
will need to work out why (possibly D3D11On12CreateDevice).