Swap Chain: Flush log file after 'Unsupported DX version' message
Normally we flush the log file every Present() call, but if we fail to
wrap the swap chain due to an unsupported DX version this will never
happen, and the message will probably never get flushed at all.

This explicitly flushes the log file after this message is printed, so
at least the reason will be known.