Overlay: Use original swap chain instead of HackerSwapChain
The overlay is supposed to bypass 3DMigoto's wrappers for rendering,
because it doesn't need them, but we were still handing it our swap
chain wrapper instead of the original swap chain, and as a result we see
our own QueryInterface calls being logged.

This also prevents the overlay from being able to use 11on12 to show in
DX12 games, since we need the IDXGISwapChain3 interface to find the
current back buffer in DX12 that the HackerSwapChain denies.