Enable overlay in DX12 games (requires Win10 SDK / vs2015 branch)
In the event that a game starts up in DX12 mode, it might be nice to use
our overlay to display a message to that effect, reminding the user to
switch to DX11. This is superior to performing a fatal exit, since the
user may need to adjust this setting from in-game, and is superior to
providing no indication that anything is wrong at all and leaving the
user wondering why nothing is working.

If the game didn't even load d3d11.dll there isn't anything we can do
(yet), but many games are loading both d3d11.dll and d3d12.dll, and
since we've hooked CreateSwapChain[ForHwnd], we will see when the game
creates the swap chain with a DX12 Command Queue, and can use that to
enable just enough functionality to get a working overlay in DX12 to
display our warning.

For this to work, we use the 11on12 interop layer in d3d11.dll - this
allows us to wrap the ID3D12Device inside an ID3D11Device, and allows us
to wrap the back buffer ID3D12Resource inside an ID3D11Texture2D,
allowing the overlay to use these as though it was DX11.

This functionality is only enabled when 3DMigoto is built with the
Windows 10 SDK (for obvious reasons) and is automatically disabled if an
older SDK is in use, and will fall back to the audible warning instead.
In practice, this means it won't show up until we migrate to the vs2015
branch.

The D3D11On12CreateDevice() function pointer is set up for our internal
use in D3D11Wrapper.cpp, but this function is not yet exported for
external games/overlays to use. We probably could add this, and 3DMigoto
would likely work with little effort for anything using this interop
layer, but it is unlikely to be used by itself, so without hooking the
rest of d3d12 we would likely be limited to a small subset of the game,
like the HUD.

The overlay itself does not yet display an error - it's just the same
overlay we know and love, but we want to add error reporting to the
overlay anyway, so this will just be another user of that when it comes.

There are a few gotchas I noted using the 11on12 interop layer:
- We have to query the current back buffer index - on DX11 this is
  always 0, but on DX12 it cycles, and we get a DXGI_ERROR_ACCESS_DENIED
  if we try to use the wrong one. This requires the IDXGISwapChain3
  interface.
- We have to call ReleaseWrappedResources() when we are done to
  transition the back buffer to the DX12 "PRESENT" state (AKA inserting
  a memory barrier to ensure the render target writes are visible), and
  ensure that 11on12 is not holding a reference to it that could
  interfere with alt+enter.
- We have to flush the DX11 immediate context before further DX12 work
  can be performed, and this must be done after all references to the
  DX11 back buffer (including unbinding it from the pipeline) has been
  released to ensure that they have really and truly been released.
- The debug layer cannot be forced when we create the DX11 device,
  because it isn't enabled on the wrapped DX12 device. It can however be
  enabled via dxcpl.exe so long as the game did not explicitly prevent
  that.
- The debug layer doesn't seem to like any choice of input state to
  CreateWrappedResource() - these are used by the 11on12 layer to
  determine which, if any, memory barrier it needs to insert before
  using the resource as a render target (since this is tracked by d3d11,
  but not d3d12), and ultimately I believe either should be okay - refer
  to the comment in the code for the reasoning.