Overlay: Add notification infrastructure
This adds infrastructure to display arbitrary messages in the overlay,
so that we can post important information for shaderhackers and end
users there.

A new LogOverlayW() helper function is added, that takes a
HackerSwapChain and uses it to retrieve the overlay (if it exists) and
posts a notification to it. This helper function will also post the
message to d3d11_log.txt so we don't need to double up.

The overlay supports several log levels - LOG_INFO, LOG_NOTICE,
LOG_WARNING and LOG_DIRE, each of which will display the message in a
different colour for some duration. LOG_INFO and LOG_NOTICE will only
show the message if hunting is currently enabled - these are intended
for messages we want a shaderhacker to see, but not an end user. These
aren't set in stone - we can tweak them as we see fit.

The overlay is currently limited to displaying no more than six
notifications at a time, which will be the most severe notifications
first, then oldest to newest. Once a message times out another message
in the queue will take its place, and a ClearNotices() function is
provided, which is intended to be called just before reloading shaders
or the d3dx.ini to clear any messages still queued from the previous
reload so the messages displayed after the reload will be relevant.