Display warning on overlay if DirectX 12 is in use (requires Win 10 SDK)
This puts our new overlay notification mechanism to use to display a
dire warning if the game is using DirectX 12, so that the user will know
what is going on and can change the game settings to use DirectX 11.

This requires that 3DMigoto is built with the Win 10 SDK, otherwise it
will fall back to using the audible warning instead.