Disassembler: Strip excess whitespace from end of lines
This has the disassembler strip excess whitespace from the end of each
line. The reason for this is that the Microsoft disassembler inserts
an extra space at the end of certain lines, such as "else ", "endif "
and "ret ", which is easy to miss when writing a regular expression to
match these lines and not at all obvious why the pattern is not
matching.

There is a risk that this may break existing fixes that have explicitly
matched these spaces, but since the ShaderRegex feature is still
relatively new and it is unusual to match these particular words in a
pattern, the risk is judged as being worth the long term benefit of
removing this gotcha.