d3dx.ini: Remove references to no_view_cache
no_view_cache was a workaround for a crash we saw in some games when
holding onto a reference of the back buffer, which we now know prevents
the buffer from being resized on alt+enter and 3DMigoto has enforced
no_view_cache for the back buffer and fake back buffers for a while now,
so remove all mentions of it.

Curious to see the comment of this in the upscaling case note that it
causes crashes (that should probably have been brought to my attention
earlier rather than worked around). I'm fairly certain that will be due
to the refcounting bug we fixed in upscaling, since the view cache would
have prevented the fake back buffer from being released hiding that bug.
Regardless, we have been enforcing no_view_cache on the fake back buffer
for a while now as well and no one complained, so it still doesn't make
sense to encourage the user to play with an option that won't do
anything, and if that wasn't the cause and the crash does still occur -
let's debug it properly.