Overlay notices: Disconnect LogOverlayW from swap chain
We want to be able to post messages to the overlay at any time, even if
we don't have a swap chain handy, such as when loading the d3dx.ini.

This disconnects the queue of notices to display from the swap chain and
overlay so that any code can post new notices to it, and the overlay
will consume messages from this queue whenever it happens to run.

As a bonus, this also reduces the line noise required to post a notice
to the overlay, and simplifies posting the DX12 warning, since that no
longer needs to be delayed until the swap chain has been created.