Overlay: Prevent notice queue being cleared multiple times / frame
The shader reload and d3dx.ini reload functionality in 3DMigoto is
separate, and one can be performed without the other. Both of these are
going to want to clear any existing messages from the overlay before
posting their own, but since they don't know about each other the config
reload would clear any messages the shader reload posts, which we don't
want.

To solve this we have the ClearNotices function refuse to run if it has
already been called in the current frame. That way whichever reload
routine runs first will clear any stale notices, but the one that runs
second won't clear any of the notices of the first.