Switch Hacker/Hooked DXGI warnings to use overlay notices
The notices in ForceDisplayParams1 indicate that the user has requested
a resolution or refresh rate override that we don't support yet. This is
probably not going to be fatal, so switch these to dire warnings in the
overlay to let the user know. This is superior to beeps, because these
options are enabled by end users who are unlikely to have turned on the
log, and wouldn't know what to do with it.

This does leave some of the audible warnings in the upscaling creation
code, because a failure here may mean we won't have an overlay to
display the warning, so a beep is better than nothing. Still switch them
to LogOverlay so if the overlay is working we will still see them.