Have profile helper use overlay notices *as well as* audible warnings
The profile helper runs very early, potentially a long time before the
game comes up and we have an overlay to display a notice, so it's a good
idea to keep the audible warnings in this case, especially as they will
tend to be triggered just after seeing the UAC prompt.

But, if we do get the game up it is nice to also display the messages on
the overlay, so log them there as well. This code is a bit special,
because part of it is only run by the helper, part is only run by
3DMigoto, and a large chunk of it is run by both, so that's another
reason we don't want to completely remove the audible warnings from here.

This does not log unrecognised lines to the overlay, because if these
have been copied out of Geforce Profile Manager they may contain corrupt
UTF16 strings, and I have no idea how DirectXTK would cope with that.