Overlay: Increase buffer size
The buffer size of 200 characters should be sufficient for most single
line messages, but since we now post HLSL compiler warnings / errors to
the overlay for custom shaders, we might have a long multi-line message
that can get cut off.

Bump the buffer size up to 1024 characters, which should be enough to
display a reasonable amount of the message, though with the large font
size we are using most of what we can see will just be the path name to
the shader.