Ini parsing / Hooking: Add notice of deprecated hooking options
If a fix that was using some of the old hooking workarounds is updated
to 1.3, these options no longer exist. Add a notification to inform the
shaderhacker they should remove them, but do so with LOG_NOTICE level so
that if an end user updates 3DMigoto they will not see the message.