Ini Parser: Un-indent all IniWarnings
The warnings are indented to try to look nicer in the log, but they look
a little weird with the new overlay notifications when they randomly
start with a couple of spaces, and the indentation was pretty
inconsistent in the log anyway. Strip all the indentation so that they
look better in the overlay and are at least consistent in the log.

If we want the indentation in the log back we could think about
IniWarning tacking it on for us, and LogOverlay stripping it
specifically for the overlay, but that's a little bit more effort and
I'm not sure it's worth it.