Shader reload: Send compile errors to the overlay
If a shader fails to compile, send the errors to the overlay for the
shaderhacker to see. If a shader compiles but had warnings we only send
them to the log file, since many of the warnings are a result of the
decompilation process and are just white noise.

This is only hooked up to shader reload, and not to initial shader
replacement for now. We could hook it up to the later, but this is
probably more of interest to shaderhackers working on and reloading a
shader live.