Notifications: Use a different font for notifications
The Courier font that we use for most of the overlay is a little too
large for notifications, and being monospace it takes up a lot of room.
Add a second sprite font from the Liberation Sans font (a free font
created by Red Hat with similar metrics to Arial) size 16 for use with
the notifications.

HLSL errors/warnings are now readable, as the path to the game no longer
fills up the entire screen width.