Set ACL on newly created directories to try to ensure we have access
Every now and then for no apparent reason the permissions on one of my
game directories gets messed up and interferes with 3DMigoto's
operation. I have no idea how or why this happens, but it usually ends
up with a bunch of googling to remember what the commands are to reset
the permissions (especially as the Windows GUI cannot cope with all
possible situations where this can occur) on the game directory to
continue.

This is a first attempt to try to have 3DMigoto cope with this itself,
by setting permissions on newly created directories to try to make sure
that it itself can access their contents. I don't think this is the only
situation I've seen - we may have to do something similar to CreateFile
to deal with cases where 3DMigoto dumps a shader that it has no
permission to read, but I'll wait until I see this happen again before
trying to sort it out.

For example, if a game directory has somehow managed to get permissions
that look like this:

    dss@Gideon ~/Subnautica
    $ icacls.exe .
    . NULL SID:(DENY)(Rc,S,REA,X,DC)
      GIDEON\dss:(F)
      NT AUTHORITY\Authenticated Users:(DENY)(W,DC)
      BUILTIN\Users:(DENY)(W,DC)
      GIDEON\None:(Rc,S,RA)
      NT AUTHORITY\Authenticated Users:(RX,W,DC)
      NT AUTHORITY\SYSTEM:(RX,W,DC)
      BUILTIN\Administrators:(RX,W,DC)
      BUILTIN\Users:(RX,W,DC)
      GIDEON\None:(DENY)(S,RD,REA,X)
      Everyone:(RX)
      NULL SID:(OI)(CI)(IO)(DENY)(Rc,S,REA,WEA,X,DC)
      CREATOR OWNER:(OI)(CI)(IO)(DENY)(S,RD,WD,AD,REA,WEA,X,DC)
      CREATOR OWNER:(OI)(CI)(IO)(D,Rc,WDAC,WO,RA,WA)
      CREATOR GROUP:(OI)(CI)(IO)(Rc,S,RA)
      NT AUTHORITY\Authenticated Users:(OI)(CI)(IO)(RX,W,DC)
      NT AUTHORITY\SYSTEM:(OI)(CI)(IO)(RX,W,DC)
      BUILTIN\Administrators:(OI)(CI)(IO)(RX,W,DC)
      BUILTIN\Users:(OI)(CI)(IO)(RX,W,DC)
      Everyone:(OI)(CI)(IO)(Rc,S,RA)

Then before this change the directories 3DMigoto creates will inherit
these permissions:

    dss@Gideon ~/Subnautica
    $ icacls.exe FrameAnalysis-2018-02-03-122406/
    FrameAnalysis-2018-02-03-122406/ NULL SID:(I)(OI)(CI)(DENY)(Rc,S,REA,WEA,X,DC)
                                     GIDEON\dss:(I)(DENY)(S,RD,WD,AD,REA,WEA,X,DC)
                                     CREATOR OWNER:(I)(OI)(CI)(IO)(DENY)(S,RD,WD,AD,REA,WEA,X,DC)
                                     GIDEON\dss:(I)(D,Rc,WDAC,WO,RA,WA)
                                     CREATOR OWNER:(I)(OI)(CI)(IO)(D,Rc,WDAC,WO,RA,WA)
                                     GIDEON\None:(I)(Rc,S,RA)
                                     CREATOR GROUP:(I)(OI)(CI)(IO)(Rc,S,RA)
                                     NT AUTHORITY\Authenticated Users:(I)(OI)(CI)(RX,W,DC)
                                     NT AUTHORITY\SYSTEM:(I)(OI)(CI)(RX,W,DC)
                                     BUILTIN\Administrators:(I)(OI)(CI)(RX,W,DC)
                                     BUILTIN\Users:(I)(OI)(CI)(RX,W,DC)
                                     Everyone:(I)(OI)(CI)(Rc,S,RA)

Which doesn't allow us to even look inside the directory:

    dss@Gideon ~/Subnautica
    $ ls FrameAnalysis-2018-02-03-122406/
    ls: cannot open directory 'FrameAnalysis-2018-02-03-122406/': Permission denied

After this change, 3DMigoto will have assigned more sensible permissions:

    dss@Gideon ~/Subnautica
    $ icacls.exe FrameAnalysis-2018-02-03-141819/
    FrameAnalysis-2018-02-03-141819/ Everyone:(OI)(CI)(RX)
                                     NT AUTHORITY\Authenticated Users:(OI)(CI)(RX,W)
                                     BUILTIN\Administrators:(OI)(CI)(F)

That actually give us access:

    dss@Gideon ~/Subnautica
    $ ls FrameAnalysis-2018-02-03-141819/
    log.txt  ShaderUsage.txt