d3dx.ini template: Disable frame analysis options by default
When frame analysis was first introduced, the main form of operation was
to dump render targets after each draw call over the entire frame, but
as time has gone on we more and more do not want to do that - either we
want it limited to part of the frame, specific shaders, or just logging.

Since dumping render targets can cost a significant amount of time in
some games, disable this by default so that the shaderhacker must
explicitly enable it if they want it. Even with no options set at all we
still get frame analysis logging.