Fix proxy loading by adding load_library_redirect setting.
Proxy loading has been broken since we added the LoadLibrary hooks,
because any calls to LoadLibrary would be forced back to the
game folder.  So any proxy DLL which would call LoadLibrary
on the System32 path would be forced back to the game
directory, which would be our d3d11.dll.  This set up an
infinite loop, because a call to CreateDevice would call us,
we'd pass on to the proxy DLL, which thought it would call
on to the System32 DLL, but in fact had our CreateDevice address,
and thus would recurse until stack death.

This fixes the problem by adding an option to skip the
LoadLibrary hooks altogether.  We've been forcing them
on all this time, but in fact, we only need them for some
games not all.  Default will remain =2, on for both d3d11.dll
and nvapi.dll both, to avoid changing behavior in this code branch.

But, this gives us the option to disable the hooks, for
scenarios that need it, like using ENB or Reshade, and probably SpecialK.

If proxy loading is activated by a named DLL, then we will
automatically disable the LoadLibrary hook because it will otherwise crash.

Two levels are available, =1 where we'll only force redirect on nvapi calls, and =2 where we'll force redirect on both nvapi and d3d11 calls.  Not positive this is necessary, but gives us some flexibility.

Logging is also improved here, to more closely describe
what it happening on dll loads.