Rework D3D11CreateDeviceAndSwapChain for SpecialK compatibility
For compatibility with Special K, we are going to undo some of our init
path simplifications - in particular, we are going to undo our
reimplementation of D3D11CreateDeviceAndSwapChain so that it calls
through to the original DX function, rather than calling
D3D11CreateDevice and CreateSwapChain. This is necessary, because
Special K has made a similar simplification to us, but has gone the
other way around calling D3D11CreateDeviceAndSwapChain from
D3D11CreateDevice, and because of the complexity of how these things are
hooked together we end up calling each other in a game of hot potato
until one of us runs out of stack.

But, we don't just want to go back to our old implementation of
D3D11CreateDeviceAndSwapChain - for one, it had a lot of duplicate code
we don't want, and two - since we now hook CreateSwapChain, DX itself
can call back into us with a device we haven't seen or wrapped yet and
we need to find a way to deal with that.

To solve this we turn our lookup_hacker_device function into
lookup_or_wrap_hacker_device so that if it is asked to look up a D3D11
device that is not yet wrapped, it will go ahead and wrap it. This way,
CreateSwapChain will then have a HackerDevice it can use to create the
HackerSwapChain. Later, when DX returns from CreateDeviceAndSwapChain we
retrieve the HackerDevice from our lookup map.

This adds helper functions to wrap the device and immediate context and
do all the interlinking and initialisation they need and to fix up the
pointers. This gives us common code we can call from both CreateDevice
calls to drastically simplify them.

Since the new CreateDeviceAndSwapChain implementation relies on DX
calling our hooked CreateSwapChain, we verify that the returned
swap chain interface is one of ours, by checking if it supports the new
IID added in the previous commit.

This also fixes refcounting issues in D3D11CreateDevice, where the
original device and context interfaces were not released after
upconverting, that might be related to a couple of crash on exit cases
seen in 1.3.