Frame Analysis: Fix infinite recursion bug when hooking is enabled
The FrameAnalysisContext held on to a copy of the original context it
was created with to allow it to call through to the original DX context
bypassing HackerContext, but when hooking is in use the original context
is modified to call back into HackerContext and we get an infinite
recursion bug and crash.

Get rid of the original context that FrameAnalysisContext holds on to,
and call GetPassThroughOrigContext1() instead when we need to bypass
HackerContext.