Extract ID3D11CreateDevice into a local routine to avoid hooks.
In SpecialK, it will hook CreateDevice, and we have a wrap on it
at the same time.  This leads to a recursive loop and stack death.  This happens because we call ID3D11CreateDevice
directly, as part of CreateDeviceAndSwapChain.

This change moves our functionality into a subroutine of
HackerCreateDevice, where it cannot be hooked.  And we
call this routine in CreateDeviceAndSwapChain to avoid
any possible hooks.

This solves the problem for SpecialK, but may not be the
best solution. Experimental.